requirements.txt
streamlit==1.31.1
pandas==2.2.0
plotly==5.18.0
sqlalchemy==2.0.25
pymysql==1.1.0
fastapi==0.109.2
uvicorn==0.27.1
pytest==7.4.3
pytest-cov==4.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic>=2.7.0
pydantic-settings>=2.9.0
python-dotenv==1.0.1
openpyxl==3.1.2
psutil==5.9.8
great-expectations==0.18.21
pyarrow==15.0.0 